+++ packages/leancode_lint/test/mock_libraries/flutter_bloc.dart
@@ -4,6 +4,7 @@ mixin MockFlutterBloc on AnalysisRuleTest {
   @override
   void setUp() {
     newPackage('flutter_bloc').addFile('lib/flutter_bloc.dart', '''
+import 'package:bloc/bloc.dart';
 import 'package:flutter/material.dart';
 
 export 'package:bloc/bloc.dart';
@@ -12,6 +13,12 @@ extension BlocContextExtention on BuildContext {
   T read<T>() => throw UnimplementedError();
 
   T watch<T>() => throw UnimplementedError();
+}
+
+class BlocProvider<T extends BlocBase<Object?>> extends Widget {
+  const BlocProvider({super.key, required this.create, this.child});
+  final T Function(BuildContext context) create;
+  final Widget? child;
 }
 ''');
     super.setUp();